Split release-manager.tsx (1142 lines) into focused modules
Extract utils/types, custom hook, and subcomponents from the oversized
release-manager component to improve maintainability and reduce context
needed when working on individual pieces.
